requirements.txt
async-timeout==4.0.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dataclasses-json==0.5.7
Deprecated==1.2.13
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
MarkupSafe==2.0.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==21.3
pyparsing==3.0.9
redis==4.3.4
requests==2.25.1
typing-inspect==0.8.0
typing_extensions==4.3.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.14.1